requirements.txt
flask=2.2.5
python=3.12.2
python-dotenv=1.0.1
requests=2.31.0
pip=23.3.1